Reorganise C++ helpers: json_helpers.hpp + unified jni_helpers.hpp
Establish a clear semantic boundary between pure JSON transforms
and JNI bridge code:

  json_helpers.hpp (new)
    Pure data-transformation functions — zero JNI, zero llama state.
    Moved from jni_server_helpers.hpp: get_result_error_message,
    results_to_json, rerank_results_to_json, build_embeddings_response_json,
    extract_first_embedding_row, parse_encoding_format,
    extract_embedding_prompt.
    Newly extracted from jllama.cpp: is_infill_request,
    parse_slot_prompt_similarity, parse_positive_int_config.
    All functions are independently unit-testable without a live JVM.

  jni_helpers.hpp (rewritten)
    Merges the former jni_helpers.hpp (Layer A: handle management) and
    jni_server_helpers.hpp (Layer B: server orchestration) into one file.
    Layer B includes json_helpers.hpp so bridge helpers can call transforms
    directly.  results_to_jstring_impl delegates to results_to_json.

  jllama.cpp
    Drops include of jni_server_helpers.hpp (superseded by jni_helpers.hpp).
    Call sites updated: *_impl suffix dropped for functions moved to
    json_helpers.hpp; requestCompletion uses is_infill_request;
    configureParallelInference uses parse_slot_prompt_similarity /
    parse_positive_int_config with a single try/catch.

  Tests
    test_json_helpers.cpp (new): ~60 tests covering all json_helpers.hpp
    functions, using fake result types — no JVM, no model required.
    test_jni_helpers.cpp: merged with relevant tests from the deleted
    test_jni_server_helpers.cpp; three fixtures cover mock JNI, server
    response queue, and JNI array access.
    test_jni_server_helpers.cpp deleted (content fully migrated).
    CMakeLists.txt: replaces test_jni_server_helpers.cpp with
    test_json_helpers.cpp in the jllama_test executable.
