--- src/test/cpp/test_jni_server_helpers.cpp
@@ -1,741 +1,0 @@
-// Tests for jni_server_helpers.hpp:
-//   - build_completion_tasks_impl
-//   - collect_task_results_impl
-//   - recv_slot_task_result_impl  (added by Finding 5)
-//
-// build_completion_tasks_impl needs:
-//   - JNIEnv  — used only for ThrowNew on the error path
-//   - server_context*  — NOT accessed when "prompt" is absent (exception thrown
-//     first), so nullptr is safe for error-path tests.
-//   - json data — provided inline in the test
-//
-// collect_task_results_impl and recv_slot_task_result_impl need:
-//   - server_response  (from server.hpp) — provides recv() / remove_waiting_task_ids()
-//   - JNIEnv           — used only for ThrowNew on the error path
-//
-// server_response is used directly: we pre-seed it with results via send()
-// before calling collect_task_results_impl().  Because recv() checks the
-// queue under a mutex+condvar, pre-seeding lets us call recv() from the same
-// thread without blocking.
-//
-// JNIEnv is mocked with the same stub technique used in test_jni_helpers.cpp:
-// a zero-filled JNINativeInterface_ table with only GetLongField (unused here)
-// and ThrowNew patched so we can observe whether an exception was raised.
-//
-// Covered scenarios:
-//   - single success result → out filled, no throw, returns true
-//   - single error result   → out empty, ThrowNew called with correct message, returns false
-//   - multiple success results → all collected in order, returns true
-//   - first result ok, second is error → cleanup, ThrowNew, returns false
-//   - waiting ids are removed on success (remove_waiting_task_ids called)
-//   - waiting ids are removed on error  (remove_waiting_task_ids called)
-
-#include <gtest/gtest.h>
-
-#include <cstring>
-#include <memory>
-#include <thread>
-#include <unordered_set>
-
-// server.hpp must come before jni_server_helpers.hpp (no include guard in server.hpp).
-#include "server.hpp"
-#include "jni_server_helpers.hpp"
-
-// ============================================================
-// Minimal concrete server_task_result subtypes for testing
-// ============================================================
-
-namespace {
-
-// A success result whose to_json() returns {"content": "<msg>"}.
-struct fake_ok_result : server_task_result {
-    std::string msg;
-    explicit fake_ok_result(int id_, std::string m) : msg(std::move(m)) { id = id_; }
-    json to_json() override { return {{"content", msg}}; }
-};
-
-// An embedding result whose to_json() returns the shape expected by
-// format_embeddings_response_oaicompat: {"embedding": [...], "tokens_evaluated": N}.
-struct fake_embedding_result : server_task_result {
-    std::vector<float> vec;
-    int tokens_evaluated;
-    explicit fake_embedding_result(int id_, std::vector<float> v, int tok = 4)
-        : vec(std::move(v)), tokens_evaluated(tok) { id = id_; }
-    json to_json() override {
-        return {{"embedding", vec}, {"tokens_evaluated", tokens_evaluated}};
-    }
-};
-
-static server_task_result_ptr make_embedding(int id_, std::vector<float> v = {0.1f, 0.2f, 0.3f}) {
-    return std::make_unique<fake_embedding_result>(id_, std::move(v));
-}
-
-// An error result — reuses the real server_task_result_error so that
-// to_json() → format_error_response() → {"message": err_msg, ...} matches
-// the exact JSON key that collect_task_results_impl reads.
-static server_task_result_ptr make_error(int id_, const std::string &msg) {
-    auto r       = std::make_unique<server_task_result_error>();
-    r->id        = id_;
-    r->err_msg   = msg;
-    r->err_type  = ERROR_TYPE_SERVER;
-    return r;
-}
-
-static server_task_result_ptr make_ok(int id_, const std::string &msg = "ok") {
-    return std::make_unique<fake_ok_result>(id_, msg);
-}
-
-// ============================================================
-// Mock JNI environment (same pattern as test_jni_helpers.cpp)
-// ============================================================
-
-static bool        g_throw_called  = false;
-static std::string g_throw_message;
-
-// NewStringUTF stub: stores the string so tests can inspect it, returns a
-// non-null sentinel so callers can distinguish success from nullptr (error).
-static std::string g_new_string_utf_value;
-static jstring     g_new_string_utf_sentinel = reinterpret_cast<jstring>(0xBEEF);
-
-static jint JNICALL stub_ThrowNew(JNIEnv *, jclass, const char *msg) {
-    g_throw_called  = true;
-    g_throw_message = msg ? msg : "";
-    return 0;
-}
-
-static jlong JNICALL stub_GetLongField(JNIEnv *, jobject, jfieldID) { return 0; }
-
-static jstring JNICALL stub_NewStringUTF(JNIEnv *, const char *utf) {
-    g_new_string_utf_value = utf ? utf : "";
-    return g_new_string_utf_sentinel;
-}
-
-JNIEnv *make_mock_env(JNINativeInterface_ &table, JNIEnv_ &env_obj) {
-    std::memset(&table, 0, sizeof(table));
-    table.ThrowNew     = stub_ThrowNew;
-    table.GetLongField = stub_GetLongField; // unused but avoids a null slot crash
-    table.NewStringUTF = stub_NewStringUTF;
-    env_obj.functions  = &table;
-    return &env_obj;
-}
-
-// Test fixture: fresh mock env + fresh server_response per test.
-struct CollectResultsFixture : ::testing::Test {
-    JNINativeInterface_ table{};
-    JNIEnv_             env_obj{};
-    JNIEnv             *env        = nullptr;
-    jclass              dummy_eclass = reinterpret_cast<jclass>(0x1);
-
-    server_response queue;
-
-    void SetUp() override {
-        env            = make_mock_env(table, env_obj);
-        g_throw_called = false;
-        g_throw_message.clear();
-        g_new_string_utf_value.clear();
-    }
-};
-
-} // namespace
-
-// ============================================================
-// Tests for get_result_error_message()
-//
-// No JNI needed — the function only calls result->to_json() and
-// performs a JSON key lookup, both of which are pure C++.
-// ============================================================
-
-TEST(GetResultErrorMessage, ErrorResult_ReturnsMessageString) {
-    server_task_result_ptr r = make_error(1, "something went wrong");
-    EXPECT_EQ(get_result_error_message(r), "something went wrong");
-}
-
-TEST(GetResultErrorMessage, DifferentMessage_ReturnsCorrectString) {
-    server_task_result_ptr r = make_error(2, "out of memory");
-    EXPECT_EQ(get_result_error_message(r), "out of memory");
-}
-
-// ============================================================
-// Single success result
-// ============================================================
-
-TEST_F(CollectResultsFixture, SingleOkResult_ReturnsTrueAndFillsOut) {
-    queue.add_waiting_task_id(1);
-    queue.send(make_ok(1, "hello"));
-
-    std::unordered_set<int> ids = {1};
-    std::vector<server_task_result_ptr> out;
-
-    bool ok = collect_task_results_impl(env, queue, ids, out, dummy_eclass);
-
-    EXPECT_TRUE(ok);
-    EXPECT_EQ(out.size(), 1u);
-    EXPECT_EQ(out[0]->to_json()["content"], "hello");
-    EXPECT_FALSE(g_throw_called);
-}
-
-// ============================================================
-// Single error result
-// ============================================================
-
-TEST_F(CollectResultsFixture, SingleErrorResult_ReturnsFalseAndThrows) {
-    queue.add_waiting_task_id(2);
-    queue.send(make_error(2, "something went wrong"));
-
-    std::unordered_set<int> ids = {2};
-    std::vector<server_task_result_ptr> out;
-
-    bool ok = collect_task_results_impl(env, queue, ids, out, dummy_eclass);
-
-    EXPECT_FALSE(ok);
-    EXPECT_TRUE(out.empty()) << "out must not be populated on error";
-    EXPECT_TRUE(g_throw_called);
-    EXPECT_EQ(g_throw_message, "something went wrong");
-}
-
-// ============================================================
-// Multiple success results
-// ============================================================
-
-TEST_F(CollectResultsFixture, MultipleOkResults_AllCollected) {
-    for (int i = 10; i < 13; ++i) {
-        queue.add_waiting_task_id(i);
-        queue.send(make_ok(i, "msg" + std::to_string(i)));
-    }
-
-    std::unordered_set<int> ids = {10, 11, 12};
-    std::vector<server_task_result_ptr> out;
-
-    bool ok = collect_task_results_impl(env, queue, ids, out, dummy_eclass);
-
-    EXPECT_TRUE(ok);
-    EXPECT_EQ(out.size(), 3u);
-    EXPECT_FALSE(g_throw_called);
-}
-
-// ============================================================
-// First ok, second error — error path cleans up remaining ids
-// ============================================================
-
-TEST_F(CollectResultsFixture, SecondResultIsError_StopsAndThrows) {
-    queue.add_waiting_task_id(20);
-    queue.add_waiting_task_id(21);
-    queue.send(make_ok(20));
-    queue.send(make_error(21, "task 21 failed"));
-
-    std::unordered_set<int> ids = {20, 21};
-    std::vector<server_task_result_ptr> out;
-
-    bool ok = collect_task_results_impl(env, queue, ids, out, dummy_eclass);
-
-    EXPECT_FALSE(ok);
-    EXPECT_TRUE(g_throw_called);
-    EXPECT_EQ(g_throw_message, "task 21 failed");
-}
-
-// ============================================================
-// Waiting ids are removed from the queue on the success path
-// ============================================================
-
-TEST_F(CollectResultsFixture, SuccessPath_WaitingIdsRemovedAfterCollect) {
-    queue.add_waiting_task_id(30);
-    queue.send(make_ok(30));
-
-    std::unordered_set<int> ids = {30};
-    std::vector<server_task_result_ptr> out;
-    (void)collect_task_results_impl(env, queue, ids, out, dummy_eclass);
-
-    // After collect, the id must no longer be in the waiting set.
-    // We verify indirectly: sending a second result for id=30 should
-    // NOT be returned by a subsequent recv for a different id — the
-    // simplest check is that waiting_task_ids no longer contains 30.
-    EXPECT_FALSE(queue.waiting_task_ids.count(30))
-        << "remove_waiting_task_ids must clear the id on success";
-}
-
-// ============================================================
-// Waiting ids are removed from the queue on the error path
-// ============================================================
-
-TEST_F(CollectResultsFixture, ErrorPath_WaitingIdsRemovedAfterError) {
-    queue.add_waiting_task_id(40);
-    queue.send(make_error(40, "err"));
-
-    std::unordered_set<int> ids = {40};
-    std::vector<server_task_result_ptr> out;
-    (void)collect_task_results_impl(env, queue, ids, out, dummy_eclass);
-
-    EXPECT_FALSE(queue.waiting_task_ids.count(40))
-        << "remove_waiting_task_ids must clear the id on error";
-}
-
-// ============================================================
-// Tests for build_completion_tasks_impl
-//
-// Only the error path is unit-testable here: the function reads
-// data["prompt"] in its own statement BEFORE accessing ctx_server, so
-// passing nullptr for ctx_server is safe when "prompt" is absent.
-//
-// The success path requires a live server_context (llama vocab + context
-// pointers) and is covered by LlamaModelTest Java integration tests.
-// ============================================================
-
-TEST_F(CollectResultsFixture, BuildTasks_MissingPrompt_ReturnsFalseAndThrows) {
-    json data = {{"n_predict", 1}}; // deliberately no "prompt" key
-    std::string completion_id = "test-cmpl-id";
-    std::vector<server_task> tasks;
-
-    // ctx_server is nullptr — safe because data.at("prompt") throws before
-    // any ctx_server member is accessed.
-    bool ok = build_completion_tasks_impl(env,
-                                          /*ctx_server=*/nullptr,
-                                          data, completion_id,
-                                          SERVER_TASK_TYPE_COMPLETION,
-                                          OAICOMPAT_TYPE_NONE,
-                                          tasks,
-                                          dummy_eclass);
-
-    EXPECT_FALSE(ok)           << "Missing 'prompt' must return false";
-    EXPECT_TRUE(g_throw_called) << "ThrowNew must be called for missing 'prompt'";
-    EXPECT_TRUE(tasks.empty()) << "tasks must remain empty on error";
-}
-
-TEST_F(CollectResultsFixture, BuildTasks_MissingPrompt_TaskTypeDoesNotAffectErrorBehaviour) {
-    // The error path is identical regardless of task_type / oaicompat.
-    // Verify INFILL behaves the same way.
-    json data = {{"input_prefix", "def f():"}, {"input_suffix", "return 1"}};
-    std::string completion_id = "infill-cmpl-id";
-    std::vector<server_task> tasks;
-
-    bool ok = build_completion_tasks_impl(env,
-                                          /*ctx_server=*/nullptr,
-                                          data, completion_id,
-                                          SERVER_TASK_TYPE_INFILL,
-                                          OAICOMPAT_TYPE_NONE,
-                                          tasks,
-                                          dummy_eclass);
-
-    EXPECT_FALSE(ok);
-    EXPECT_TRUE(g_throw_called);
-    EXPECT_TRUE(tasks.empty());
-}
-
-// ============================================================
-// Tests for recv_slot_task_result_impl
-//
-// Pre-seed the server_response queue (same technique as collect tests):
-// calling send() before recv() satisfies the condvar predicate immediately,
-// so recv() returns without blocking even from the same thread.
-//
-// NewStringUTF is stubbed to return a sentinel non-null jstring and capture
-// the serialised JSON so we can verify the success path.
-// ============================================================
-
-TEST_F(CollectResultsFixture, RecvSlotResult_SuccessResult_ReturnsNonNullAndDoesNotThrow) {
-    queue.add_waiting_task_id(50);
-    queue.send(make_ok(50, "slot-ok"));
-
-    jstring result = recv_slot_task_result_impl(env, queue, 50, dummy_eclass);
-
-    EXPECT_NE(result, nullptr)       << "success result must return non-null jstring";
-    EXPECT_FALSE(g_throw_called)     << "ThrowNew must not be called on success";
-    // The stub captures the serialised JSON passed to NewStringUTF.
-    EXPECT_FALSE(g_new_string_utf_value.empty())
-        << "NewStringUTF must be called with the result JSON";
-    EXPECT_NE(g_new_string_utf_value.find("slot-ok"), std::string::npos)
-        << "result JSON must contain the content from the fake result";
-}
-
-TEST_F(CollectResultsFixture, RecvSlotResult_ErrorResult_ReturnsNullAndThrows) {
-    queue.add_waiting_task_id(51);
-    queue.send(make_error(51, "slot operation failed"));
-
-    jstring result = recv_slot_task_result_impl(env, queue, 51, dummy_eclass);
-
-    EXPECT_EQ(result, nullptr)      << "error result must return nullptr";
-    EXPECT_TRUE(g_throw_called)     << "ThrowNew must be called on error";
-    EXPECT_EQ(g_throw_message, "slot operation failed");
-}
-
-TEST_F(CollectResultsFixture, RecvSlotResult_WaitingIdRemovedAfterSuccess) {
-    queue.add_waiting_task_id(52);
-    queue.send(make_ok(52));
-
-    (void)recv_slot_task_result_impl(env, queue, 52, dummy_eclass);
-
-    EXPECT_FALSE(queue.waiting_task_ids.count(52))
-        << "remove_waiting_task_id must clear the id on success";
-}
-
-TEST_F(CollectResultsFixture, RecvSlotResult_WaitingIdRemovedAfterError) {
-    queue.add_waiting_task_id(53);
-    queue.send(make_error(53, "err"));
-
-    (void)recv_slot_task_result_impl(env, queue, 53, dummy_eclass);
-
-    EXPECT_FALSE(queue.waiting_task_ids.count(53))
-        << "remove_waiting_task_id must clear the id on error";
-}
-
-// ============================================================
-// Tests for results_to_jstring_impl
-//
-// Verifies that the serialisation helper produces the right shape:
-//   - single result  → bare JSON object
-//   - multiple results → JSON array
-//
-// NewStringUTF is stubbed via the fixture's mock env; g_new_string_utf_value
-// captures whatever string was passed to it.
-// ============================================================
-
-TEST_F(CollectResultsFixture, ResultsToJstring_SingleResult_ReturnsBareObject) {
-    std::vector<server_task_result_ptr> results;
-    results.push_back(make_ok(1, "hello"));
-
-    jstring js = results_to_jstring_impl(env, results);
-
-    EXPECT_NE(js, nullptr);
-    EXPECT_FALSE(g_new_string_utf_value.empty());
-
-    // The top-level JSON must be an object (not an array).
-    json parsed = json::parse(g_new_string_utf_value);
-    EXPECT_TRUE(parsed.is_object())
-        << "single result must serialise as a bare JSON object, got: "
-        << g_new_string_utf_value;
-    EXPECT_EQ(parsed.value("content", ""), "hello");
-}
-
-TEST_F(CollectResultsFixture, ResultsToJstring_MultipleResults_ReturnsArray) {
-    std::vector<server_task_result_ptr> results;
-    results.push_back(make_ok(2, "first"));
-    results.push_back(make_ok(3, "second"));
-
-    jstring js = results_to_jstring_impl(env, results);
-
-    EXPECT_NE(js, nullptr);
-    json parsed = json::parse(g_new_string_utf_value);
-    EXPECT_TRUE(parsed.is_array())
-        << "multiple results must serialise as a JSON array, got: "
-        << g_new_string_utf_value;
-    ASSERT_EQ(parsed.size(), 2u);
-    EXPECT_EQ(parsed[0].value("content", ""), "first");
-    EXPECT_EQ(parsed[1].value("content", ""), "second");
-}
-
-TEST_F(CollectResultsFixture, ResultsToJstring_EmptyVector_ReturnsEmptyArray) {
-    std::vector<server_task_result_ptr> results;
-
-    jstring js = results_to_jstring_impl(env, results);
-
-    EXPECT_NE(js, nullptr);
-    json parsed = json::parse(g_new_string_utf_value);
-    EXPECT_TRUE(parsed.is_array());
-    EXPECT_TRUE(parsed.empty());
-}
-
-// ============================================================
-// Tests for results_to_json_impl
-//
-// Pure-logic counterpart to results_to_jstring_impl — no JNI needed.
-// ============================================================
-
-TEST(ResultsToJsonImpl, SingleResult_ReturnsObjectDirectly) {
-    std::vector<server_task_result_ptr> results;
-    results.push_back(make_ok(1, "only"));
-
-    json out = results_to_json_impl(results);
-
-    EXPECT_TRUE(out.is_object());
-    EXPECT_EQ(out.value("content", ""), "only");
-}
-
-TEST(ResultsToJsonImpl, MultipleResults_ReturnsArray) {
-    std::vector<server_task_result_ptr> results;
-    results.push_back(make_ok(1, "a"));
-    results.push_back(make_ok(2, "b"));
-
-    json out = results_to_json_impl(results);
-
-    EXPECT_TRUE(out.is_array());
-    ASSERT_EQ(out.size(), 2u);
-    EXPECT_EQ(out[0].value("content", ""), "a");
-    EXPECT_EQ(out[1].value("content", ""), "b");
-}
-
-TEST(ResultsToJsonImpl, EmptyVector_ReturnsEmptyArray) {
-    std::vector<server_task_result_ptr> results;
-
-    json out = results_to_json_impl(results);
-
-    EXPECT_TRUE(out.is_array());
-    EXPECT_TRUE(out.empty());
-}
-
-// ============================================================
-// Tests for json_to_jstring_impl
-//
-// Verifies that any json value is serialised correctly via
-// dump() + NewStringUTF.  The stub captures the string so tests
-// can round-trip parse it.
-// ============================================================
-
-TEST_F(CollectResultsFixture, JsonToJstring_Object_RoundTrips) {
-    json j = {{"key", "value"}, {"n", 42}};
-
-    jstring js = json_to_jstring_impl(env, j);
-
-    EXPECT_NE(js, nullptr);
-    EXPECT_FALSE(g_new_string_utf_value.empty());
-    json parsed = json::parse(g_new_string_utf_value);
-    EXPECT_TRUE(parsed.is_object());
-    EXPECT_EQ(parsed.value("key", ""), "value");
-    EXPECT_EQ(parsed.value("n", 0), 42);
-}
-
-TEST_F(CollectResultsFixture, JsonToJstring_Array_RoundTrips) {
-    json j = json::array({1, 2, 3});
-
-    jstring js = json_to_jstring_impl(env, j);
-
-    EXPECT_NE(js, nullptr);
-    json parsed = json::parse(g_new_string_utf_value);
-    EXPECT_TRUE(parsed.is_array());
-    ASSERT_EQ(parsed.size(), 3u);
-    EXPECT_EQ(parsed[0], 1);
-    EXPECT_EQ(parsed[2], 3);
-}
-
-TEST_F(CollectResultsFixture, JsonToJstring_ReturnsSentinelFromStub) {
-    // The mock NewStringUTF returns the 0xBEEF sentinel — verify the
-    // function propagates it unchanged so callers get a non-null jstring.
-    json j = {{"ok", true}};
-
-    jstring js = json_to_jstring_impl(env, j);
-
-    EXPECT_EQ(js, reinterpret_cast<jstring>(0xBEEF));
-}
-
-// ============================================================
-// Tests for extract_first_embedding_row
-//
-// Pure computation — no JNI or llama context needed.
-// ============================================================
-
-TEST(ExtractFirstEmbeddingRow, SingleRow_ReturnsRow) {
-    json j = {{"embedding", {{0.1f, 0.2f, 0.3f}}}};
-    auto row = extract_first_embedding_row(j);
-    ASSERT_EQ(row.size(), 3u);
-    EXPECT_FLOAT_EQ(row[0], 0.1f);
-    EXPECT_FLOAT_EQ(row[1], 0.2f);
-    EXPECT_FLOAT_EQ(row[2], 0.3f);
-}
-
-TEST(ExtractFirstEmbeddingRow, MultipleRows_ReturnsFirstRowOnly) {
-    json j = {{"embedding", {{1.0f, 2.0f}, {3.0f, 4.0f}, {5.0f, 6.0f}}}};
-    auto row = extract_first_embedding_row(j);
-    ASSERT_EQ(row.size(), 2u);
-    EXPECT_FLOAT_EQ(row[0], 1.0f);
-    EXPECT_FLOAT_EQ(row[1], 2.0f);
-}
-
-TEST(ExtractFirstEmbeddingRow, MissingEmbeddingKey_ThrowsJsonException) {
-    json j = {{"other_key", "value"}};
-    EXPECT_THROW(extract_first_embedding_row(j), nlohmann::json::exception);
-}
-
-TEST(ExtractFirstEmbeddingRow, EmptyOuterArray_ThrowsRuntimeError) {
-    json j = {{"embedding", json::array()}};
-    EXPECT_THROW(extract_first_embedding_row(j), std::runtime_error);
-}
-
-TEST(ExtractFirstEmbeddingRow, EmptyInnerArray_ThrowsRuntimeError) {
-    json j = {{"embedding", {json::array()}}};
-    EXPECT_THROW(extract_first_embedding_row(j), std::runtime_error);
-}
-
-TEST(ExtractFirstEmbeddingRow, LargeRow_AllValuesPreserved) {
-    std::vector<float> vals(128);
-    for (int i = 0; i < 128; ++i) vals[i] = static_cast<float>(i) * 0.01f;
-    json j = {{"embedding", {vals}}};
-    auto row = extract_first_embedding_row(j);
-    ASSERT_EQ(row.size(), 128u);
-    for (int i = 0; i < 128; ++i) {
-        EXPECT_FLOAT_EQ(row[i], static_cast<float>(i) * 0.01f);
-    }
-}
-
-// ============================================================
-// Tests for parse_encoding_format_impl
-//
-// Pure computation — no JNI or llama context needed.
-// ============================================================
-
-TEST(ParseEncodingFormat, FieldAbsent_ReturnsFalse) {
-    json body = {{"model", "text-embedding-ada-002"}};
-    EXPECT_FALSE(parse_encoding_format_impl(body));
-}
-
-TEST(ParseEncodingFormat, ExplicitFloat_ReturnsFalse) {
-    json body = {{"encoding_format", "float"}};
-    EXPECT_FALSE(parse_encoding_format_impl(body));
-}
-
-TEST(ParseEncodingFormat, Base64_ReturnsTrue) {
-    json body = {{"encoding_format", "base64"}};
-    EXPECT_TRUE(parse_encoding_format_impl(body));
-}
-
-TEST(ParseEncodingFormat, UnknownFormat_ThrowsInvalidArgument) {
-    json body = {{"encoding_format", "binary"}};
-    EXPECT_THROW(parse_encoding_format_impl(body), std::invalid_argument);
-}
-
-TEST(ParseEncodingFormat, EmptyString_ThrowsInvalidArgument) {
-    json body = {{"encoding_format", ""}};
-    EXPECT_THROW(parse_encoding_format_impl(body), std::invalid_argument);
-}
-
-TEST(ParseEncodingFormat, UnknownFormat_MessageMentionsValidOptions) {
-    json body = {{"encoding_format", "hex"}};
-    try {
-        parse_encoding_format_impl(body);
-        FAIL() << "Expected std::invalid_argument";
-    } catch (const std::invalid_argument &e) {
-        EXPECT_NE(std::string(e.what()).find("float"), std::string::npos)
-            << "error message should mention \"float\"";
-        EXPECT_NE(std::string(e.what()).find("base64"), std::string::npos)
-            << "error message should mention \"base64\"";
-    }
-}
-
-// ============================================================
-// Tests for extract_embedding_prompt_impl
-//
-// Pure computation — no JNI or llama context needed.
-// ============================================================
-
-TEST(ExtractEmbeddingPrompt, InputKey_ReturnsValueAndDoesNotSetFlag) {
-    bool flag = true; // pre-set to verify it gets cleared
-    json body = {{"input", "hello world"}};
-    json prompt = extract_embedding_prompt_impl(body, flag);
-    EXPECT_EQ(prompt, "hello world");
-    EXPECT_FALSE(flag) << "force_no_oaicompat must be false when \"input\" is used";
-}
-
-TEST(ExtractEmbeddingPrompt, ContentKey_ReturnsValueAndSetsFlag) {
-    bool flag = false;
-    json body = {{"content", "some text"}};
-    json prompt = extract_embedding_prompt_impl(body, flag);
-    EXPECT_EQ(prompt, "some text");
-    EXPECT_TRUE(flag) << "force_no_oaicompat must be true when \"content\" is used";
-}
-
-TEST(ExtractEmbeddingPrompt, InputTakesPriorityOverContent) {
-    bool flag = false;
-    json body = {{"input", "from input"}, {"content", "from content"}};
-    json prompt = extract_embedding_prompt_impl(body, flag);
-    EXPECT_EQ(prompt, "from input");
-    EXPECT_FALSE(flag) << "\"input\" path must not set force_no_oaicompat";
-}
-
-TEST(ExtractEmbeddingPrompt, NeitherKey_ThrowsInvalidArgument) {
-    bool flag = false;
-    json body = {{"model", "text-embedding-ada-002"}};
-    EXPECT_THROW(extract_embedding_prompt_impl(body, flag), std::invalid_argument);
-}
-
-TEST(ExtractEmbeddingPrompt, EmptyBody_ThrowsInvalidArgument) {
-    bool flag = false;
-    EXPECT_THROW(extract_embedding_prompt_impl(json::object(), flag), std::invalid_argument);
-}
-
-TEST(ExtractEmbeddingPrompt, ArrayPrompt_ReturnedAsIs) {
-    // "input" may be an array of strings (batch embedding); the function must
-    // return the JSON value unchanged without trying to coerce it.
-    bool flag = false;
-    json body = {{"input", {"sentence one", "sentence two"}}};
-    json prompt = extract_embedding_prompt_impl(body, flag);
-    ASSERT_TRUE(prompt.is_array());
-    ASSERT_EQ(prompt.size(), 2u);
-    EXPECT_EQ(prompt[0], "sentence one");
-    EXPECT_EQ(prompt[1], "sentence two");
-    EXPECT_FALSE(flag);
-}
-
-// ============================================================
-// Tests for build_embeddings_response_json_impl
-//
-// Pure computation — no JNI or llama context needed.
-// Uses fake_embedding_result (added above) for OAI-path tests that
-// need "embedding" + "tokens_evaluated" in the result JSON.
-// ============================================================
-
-TEST(BuildEmbeddingsResponseJson, NonOai_SingleResult_ReturnsBareArray) {
-    std::vector<server_task_result_ptr> results;
-    results.push_back(make_embedding(1, {0.1f, 0.2f}));
-
-    json out = build_embeddings_response_json_impl(results, json::object(),
-                                                    OAICOMPAT_TYPE_NONE, false);
-
-    ASSERT_TRUE(out.is_array());
-    ASSERT_EQ(out.size(), 1u);
-    EXPECT_TRUE(out[0].contains("embedding"));
-}
-
-TEST(BuildEmbeddingsResponseJson, NonOai_MultipleResults_AllInArray) {
-    std::vector<server_task_result_ptr> results;
-    results.push_back(make_embedding(1, {0.1f}));
-    results.push_back(make_embedding(2, {0.2f}));
-    results.push_back(make_embedding(3, {0.3f}));
-
-    json out = build_embeddings_response_json_impl(results, json::object(),
-                                                    OAICOMPAT_TYPE_NONE, false);
-
-    ASSERT_TRUE(out.is_array());
-    EXPECT_EQ(out.size(), 3u);
-}
-
-TEST(BuildEmbeddingsResponseJson, OaiFloat_WrapsWithOaiStructure) {
-    std::vector<server_task_result_ptr> results;
-    results.push_back(make_embedding(1, {0.5f, 0.6f, 0.7f}));
-
-    json body = {{"model", "text-embedding-ada-002"}};
-    json out = build_embeddings_response_json_impl(results, body,
-                                                    OAICOMPAT_TYPE_EMBEDDING, false);
-
-    EXPECT_TRUE(out.is_object())          << "OAI response must be an object";
-    EXPECT_EQ(out.value("object", ""),   "list");
-    EXPECT_TRUE(out.contains("data"))    << "OAI response must have \"data\"";
-    EXPECT_TRUE(out.contains("usage"))   << "OAI response must have \"usage\"";
-    EXPECT_EQ(out.value("model", ""),    "text-embedding-ada-002");
-
-    ASSERT_TRUE(out["data"].is_array());
-    ASSERT_EQ(out["data"].size(), 1u);
-    EXPECT_EQ(out["data"][0].value("object", ""), "embedding");
-}
-
-TEST(BuildEmbeddingsResponseJson, OaiBase64_EmbeddingEncodedAsString) {
-    std::vector<server_task_result_ptr> results;
-    results.push_back(make_embedding(1, {1.0f, 2.0f}));
-
-    json out = build_embeddings_response_json_impl(results, json::object(),
-                                                    OAICOMPAT_TYPE_EMBEDDING, /*use_base64=*/true);
-
-    ASSERT_TRUE(out["data"].is_array());
-    ASSERT_EQ(out["data"].size(), 1u);
-    // base64 path stores embedding as a string, not an array
-    EXPECT_TRUE(out["data"][0]["embedding"].is_string())
-        << "base64 embedding must be serialised as a string";
-}
-
-TEST(BuildEmbeddingsResponseJson, OaiUsage_TokensSummedAcrossResults) {
-    std::vector<server_task_result_ptr> results;
-    results.push_back(std::make_unique<fake_embedding_result>(1, std::vector<float>{0.1f}, /*tok=*/3));
-    results.push_back(std::make_unique<fake_embedding_result>(2, std::vector<float>{0.2f}, /*tok=*/5));
-
-    json out = build_embeddings_response_json_impl(results, json::object(),
-                                                    OAICOMPAT_TYPE_EMBEDDING, false);
-
-    EXPECT_EQ(out["usage"].value("prompt_tokens", 0), 8)
-        << "usage.prompt_tokens must be sum of tokens_evaluated across all results";
-}
